Get add_rtree_query to run in Visual Studion 12 - In eigenverb_collection::query_rtree, change the type of   "interface" argument from "int" to "size_t". - Remove #include <sys/time.h> from dead_reckon_test.cc.  Not used   and not found in Windows.

+++ eigenverb/eigenverb_collection.cc
@@ -43,7 +43,7 @@ box eigenverb_collection::build_box(eigenverb verb, int sigma)
  * spatial box specified the rcv_eigenverb.
  * Results are return via the third parameter.
  */
-void eigenverb_collection::query_rtree(int interface, eigenverb verb,
+void eigenverb_collection::query_rtree(size_t interface, eigenverb verb,
 											  std::vector<value_pair>& result_s)
 {
 	int scaling = 1;
